+++ api/src/test/java/org/pmiops/workbench/actionaudit/ActionAuditQueryServiceTest.java
@@ -3,7 +3,7 @@ package org.pmiops.workbench.actionaudit;
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
-import static org.pmiops.workbench.utils.TemporalAssertions.assertTimeWithinTolerance;
+import static org.pmiops.workbench.utils.TimeAssertions.assertTimeWithinTolerance;
 
 import com.google.api.gax.paging.Page;
 import com.google.cloud.bigquery.EmptyTableResult;
