+++ api/src/test/java/org/pmiops/workbench/utils/mappers/AuditLogEntryMapperTest.java
@@ -2,7 +2,7 @@ package org.pmiops.workbench.utils.mappers;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth8.assertThat;
-import static org.pmiops.workbench.utils.TemporalAssertions.assertTimeWithinTolerance;
+import static org.pmiops.workbench.utils.TimeAssertions.assertTimeWithinTolerance;
 
 import com.google.common.collect.ImmutableList;
 import java.time.OffsetDateTime;
